+++ ephios/extra/management/commands/send_notifications.py
@@ -1,6 +1,6 @@
 from django.core.management import BaseCommand
 
-from ephios.core.notifications.backends import send_all_notifications
+from ephios.core.services.notifications.backends import send_all_notifications
 
 
 class Command(BaseCommand):
